Setup global setup.js file using setupFilesAfterEnv * add to jest.config.js setupFilesAfterEnv * remove useless import and mock code from tests * fix tests
