+++ tests/views/detailRecord.spec.js
@@ -1,48 +1,6 @@
 import DetailRecord from "@/views/DetailRecord";
-import {createLocalVue, mount} from "@vue/test-utils";
-
-import Vuetify from "vuetify";
-import Vue from "vue";
-
-Vue.use(Vuetify)
-
-import Vuex from "vuex";
-import VueMeta from "vue-meta";
-import titlePlugin from "@/utils/title.plugin"
 import store from "@/store"
-import dateFilter from "@/filters/date.filter";
-import localizeFilter from "@/filters/localize.filter";
-import numberFilter from "@/filters/number.filter";
-
-const vuetify = new Vuetify()
-const localVue = createLocalVue()
-localVue.use(Vuex)
-localVue.use(VueMeta)
-localVue.use(titlePlugin)
-localVue.filter('localize', localizeFilter)
-localVue.filter('number', numberFilter)
-localVue.filter('date', dateFilter)
-
-Storage.prototype.getItem = jest.fn(
-  key => key === 'locale' && 'ru-RU'
-)
-const mockOnceVal = jest.fn()
-jest.mock('firebase/app', () => ({
-    auth: jest.fn().mockReturnThis(),
-    currentUser: {
-      email: 'user@example.com', uid: 123, emailVerified: true
-    },
-    initializeApp: jest.fn(),
-    database: () => ({
-      ref: jest.fn(() => ({
-        child: jest.fn().mockReturnThis(),
-        once: () => ({
-          val: mockOnceVal
-        })
-      }))
-    })
-  })
-);
+import {mount} from "@vue/test-utils";
 
 describe('DetailRecord.vue view component testsuite', () => {
   let wrapper
